requirements.txt
numpy>=1.18.0
opencv-python>=4.5.0
pyyaml>=5.3.1
tqdm>=4.41.0
matplotlib>=3.2.0
scipy>=1.4.1
hailo-ai>=4.12.0
pillow>=7.1.2
requests>=2.23.0
protobuf>=3.11.0 